ASCII continuous convertion

+++ test/Utility.h
@@ -74,9 +74,9 @@ void print_clock(char* s,clock_t str, clock_t end){
     printf("%s Time %f", s,total_t);
 }
 
-static int power_int(int x, int n){
-    int p = 1;
-    for(int i=0;i<n;i++){
+static long power_int(int x, int n){
+    long p = 1;
+    for(long i=0;i<n;i++){
         p *= x;
     }
     return p;
@@ -86,12 +86,12 @@ static int power_int(int x, int n){
 
 static int HashString(char s[], int D, int Q){
     int len = strlen(s);
-    int hash = 0;
+    long hash = 0;
     
-    for(int i=0;i<len;i++){
-        hash += s[i]*power_int(D, len-i-1);
+    for(long i=0;i<len;i++){
+        hash += (s[i]*power_int(D, len-i-1)) ;
     }
-    return hash% Q;
+    return (int)(hash % Q);
 };
 
 #endif
